Gather server routed together with server and move shared components
