+++ distributed/dashboard/components/worker.py
@@ -27,9 +27,9 @@ from distributed.dashboard.components.shared import (
     DashboardComponent,
     ProfileTimePlot,
     ProfileServer,
+    SystemMonitor,
 )
 from distributed.dashboard.utils import transpose, without_property_validation
-from distributed.compatibility import WINDOWS
 from distributed.diagnostics.progress_stream import color_of
 from distributed.metrics import time
 from distributed.utils import log_errors, key_split, format_time
@@ -51,7 +51,6 @@ BOKEH_THEME = Theme(os.path.join(os.path.dirname(__file__), "..", "theme.yaml"))
 template_variables = {"pages": ["status", "system", "profile", "crossfilter"]}
 
 
-# TODO Move DashboardComponents into components submodule
 class StateTable(DashboardComponent):
     """ Currently running tasks """
 
@@ -427,96 +426,6 @@ class CrossFilter(DashboardComponent):
         except Exception as e:
             logger.exception(e)
             raise
-
-
-class SystemMonitor(DashboardComponent):
-    def __init__(self, worker, height=150, **kwargs):
-        self.worker = worker
-
-        names = worker.monitor.quantities
-        self.last = 0
-        self.source = ColumnDataSource({name: [] for name in names})
-        self.source.data.update(self.get_data())
-
-        x_range = DataRange1d(follow="end", follow_interval=20000, range_padding=0)
-
-        tools = "reset,xpan,xwheel_zoom"
-
-        self.cpu = figure(
-            title="CPU",
-            x_axis_type="datetime",
-            height=height,
-            tools=tools,
-            x_range=x_range,
-            **kwargs
-        )
-        self.cpu.line(source=self.source, x="time", y="cpu")
-        self.cpu.yaxis.axis_label = "Percentage"
-        self.mem = figure(
-            title="Memory",
-            x_axis_type="datetime",
-            height=height,
-            tools=tools,
-            x_range=x_range,
-            **kwargs
-        )
-        self.mem.line(source=self.source, x="time", y="memory")
-        self.mem.yaxis.axis_label = "Bytes"
-        self.bandwidth = figure(
-            title="Bandwidth",
-            x_axis_type="datetime",
-            height=height,
-            x_range=x_range,
-            tools=tools,
-            **kwargs
-        )
-        self.bandwidth.line(source=self.source, x="time", y="read_bytes", color="red")
-        self.bandwidth.line(source=self.source, x="time", y="write_bytes", color="blue")
-        self.bandwidth.yaxis.axis_label = "Bytes / second"
-
-        # self.cpu.yaxis[0].formatter = NumeralTickFormatter(format='0%')
-        self.bandwidth.yaxis[0].formatter = NumeralTickFormatter(format="0.0b")
-        self.mem.yaxis[0].formatter = NumeralTickFormatter(format="0.0b")
-
-        plots = [self.cpu, self.mem, self.bandwidth]
-
-        if not WINDOWS:
-            self.num_fds = figure(
-                title="Number of File Descriptors",
-                x_axis_type="datetime",
-                height=height,
-                x_range=x_range,
-                tools=tools,
-                **kwargs
-            )
-
-            self.num_fds.line(source=self.source, x="time", y="num_fds")
-            plots.append(self.num_fds)
-
-        if "sizing_mode" in kwargs:
-            kw = {"sizing_mode": kwargs["sizing_mode"]}
-        else:
-            kw = {}
-
-        if not WINDOWS:
-            self.num_fds.y_range.start = 0
-        self.mem.y_range.start = 0
-        self.cpu.y_range.start = 0
-        self.bandwidth.y_range.start = 0
-
-        self.root = column(*plots, **kw)
-        self.worker.monitor.update()
-
-    def get_data(self):
-        d = self.worker.monitor.range_query(start=self.last)
-        d["time"] = [x * 1000 for x in d["time"]]
-        self.last = self.worker.monitor.count
-        return d
-
-    @without_property_validation
-    def update(self):
-        with log_errors():
-            self.source.stream(self.get_data(), 1000)
 
 
 class Counters(DashboardComponent):
